Don't use and dispose of a new HttpClient for each log
If a new HttpClient is used and disposed of for each log, then we will
end up with port starvation on the machine that is running the app using
HttpEndpointLogProvider. However, if we only ever create one HttpClient,
then the app will bypass load balancing and DNS changes. The middle
ground is to use an HttpClient for some amount of time, then use another
one for some time, then use another one, et cetera. Each time we swap
out for a new HttpClient, we dispose of the one that had been switched
out during the previous swap. That way, any requests that are in process
have a time to finish.