Improve serialization of state objects coming from asp.net core
One common type that renders poorly implements IEnumerable<string, object> but is not an IDictionary<string, object>. This causes json.net to serialize it as an array of Key/Value objects - very ugly. This change looks for state objects that are basically a dictionary and converts them to an actual dictionary.