Rename Providers to LogProviders
The name "Providers" is not descriptive enough when we have two kinds of them - "Log" and "Context", so this commit renames it to "LogParameters". However, troubleshooting configuration errors when different versions have different names is very difficult. In order to eliminate that problem, we want to support either "LogProviders" (preferred) or "Providers" (legacy). RockLib.Configuration.ObjectFactory version 1.6.0 solves that problem by allowing one or more alternate names to be assigned to a member by decorating it with an `[AlternateName]` attribute. The `Logger` constructor parameter does that.