Don't register for domain events in BackgroundLogProcessor constructor
The only BackgroundLogProcessor that we need to "try really, really hard to dispose" is the one used by the ProcessingMode constructor, so don't register domain events for every instance. Instead, use same instance of BackgroundLogProcessor for each Logger with ProcessingMode.Background, and only register domain events to dispose this singleton.